requirements.txt
discord.py
aiohttp
matplotlib
